+++ Zip/ExtraFieldHeader.swift
@@ -1,0 +1,4 @@
+struct ExtraFieldHeader {
+  let id: ZipExtraField.ID
+  let dataSize: UInt16
+}
